pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adaptive-retrieval-qa-with-answerability-calibration"
version = "0.1.0"
description = "A retrieval-augmented question answering system with confidence calibration for answerability prediction"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Alireza Shojaei"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "datasets>=2.12.0",
    "evaluate>=0.4.0",
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "scikit-learn>=1.2.0",
    "mlflow>=2.4.0",
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
    "faiss-cpu>=1.7.0",
    "nltk>=3.8",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=5.0",
    "mypy>=1.0",
    "jupyter>=1.0.0",
]

[project.urls]
Repository = "https://github.com/A-SHOJAEI/adaptive-retrieval-qa-with-answerability-calibration"

[tool.hatch.build.targets.wheel]
packages = ["src/adaptive_retrieval_qa_with_answerability_calibration"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src --cov-report=html --cov-report=term-missing"